change password is done, including web service side and UI side
