+++ IHNServer/WebRoot/index-origin.html
@@ -1,0 +1,31 @@
+<!doctype html>
+<html class="no-js">
+	<head>
+		<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
+		<meta http-equiv="X-UA-Compatible" content="IE=EDGE">
+		<title>Inside Home Navigator</title>
+		<link rel="stylesheet" href="styles/bootstrap.css">
+		<link rel="stylesheet" href="styles/app.css">
+	</head>
+	
+	<body>
+		<div id="ihn-header-container">
+	  		<div id="ihn-header"></div>
+		</div>
+		<div id="ihn-maincontent-container">
+		   <div id="ihn-sidemenu"></div>
+		   <div id="ihn-maincontent">
+		   </div>
+		</div>
+		<div id="ihn-footer-container">
+		   <div id="ihn-footer"></div>
+		</div>
+		
+		<script>
+		var require = {
+			baseUrl: 'scripts/app/property'
+		};
+		</script>
+		<script data-main="../../main" src="scripts/libs/require/require.js"></script>
+	</body>
+</html>
