+++ IHNServer/WebRoot/index.html
@@ -1,24 +1,21 @@
 <!doctype html>
-<html class="no-js">
+<html>
 	<head>
 		<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
 		<meta http-equiv="X-UA-Compatible" content="IE=EDGE">
-		<title>Inside Home Navigator</title>
+		<title>Inside Home Navigator - Find my car</title>
 		<link rel="stylesheet" href="styles/bootstrap.css">
 		<link rel="stylesheet" href="styles/app.css">
+		<link rel="stylesheet" href="styles/jquery-ui.css">
 	</head>
 	
 	<body>
-		<div id="ihn-header-container">
-	  		<div id="ihn-header"></div>
+		<div id="ihn-banner"></div>
+		<hr>
+		<div id="ihn-maincontent" style="margin:5px 250px 5px 10px;border:1px solid gray">
 		</div>
-		<div id="ihn-maincontent-container">
-		   <div id="ihn-sidemenu"></div>
-		   <div id="ihn-maincontent">
-		   </div>
-		</div>
-		<div id="ihn-footer-container">
-		   <div id="ihn-footer"></div>
+		<hr>
+		<div id="ihn-footer" style="margin:5px 250px 5px 10px; border:0px">
 		</div>
 		
 		<script>
