--- IHNServer/WebRoot/scripts/app/property/collections/ihn/sidemenus-collection.js
@@ -1,9 +1,0 @@
-define([
-], function() {
-  
-  var SideMenuCollection = Backbone.Collection.extend({
-  });
-
-  return SideMenuCollection;
-  
-});
