+++ IHNServer/WebRoot/scripts/app/property/routers/ihn-router.js
@@ -2,47 +2,44 @@ define([
     'jquery',
     'underscore',
     'backbone',
-    'views/ihn/demo/demo-view'
-], function($, _, Backbone, DemoView) {
+    'views/ihn/maincontent-view'
+], function($, _, Backbone, MainContentView) {
 
   var slice = Array.prototype.slice;
   
   var AppRouter = Backbone.Router.extend({
     
     curRoute: 'demo',
-    headerView: null,
-    sideMenuView: null,
+    bannerView: null,
     mainContentView: null,
     footerView: null,
     
     initialize: function(options) {
-      this.headerView = options.headerView;
-      this.sideMenuView = options.sideMenuView;
+      this.bannerView = options.bannerView;
       this.mainContentView = options.mainContentView;
       this.footerView = options.footerView;
+      
+      this.showMapPage(null);
     },
     
-    routes: {
-      'demo': 'showDemoPage'
-    },
-    
-    showDemoPage: function(href) {
-      curRoute = 'demo';
-      this.show(DemoView);
+    showMapPage: function(href) {
+      curRoute = 'map';
+      this.show(this.mainContentView);
     },
     
     show: function(View, options) {
       var args = slice.call(arguments, _.isObject(options) ? 2 : 1);
-      this.renderContent(View, args);
+      //this.renderContent(View, args);
     },
     
     renderContent: function(ContentView, args){
-      var main = this.mainContentView, $mainContainer = main.$el;
-      var view = new ContentView({'linkOptions': args});
-      $mainContainer.html(view.render().$el);
+      //var main = this.mainContentView;
+      //var view = new ContentView({'linkOptions': args});
+      //$mainContainer.html(view.render().$el);
     }
     
   });
   
   return AppRouter;
+  
 });
