+++ IHNServer/WebRoot/scripts/app/property/templates/ihn/banner-template.html
