--- IHNServer/WebRoot/scripts/app/property/templates/ihn/demo/demo-template.html
@@ -1,3 +1,0 @@
-<canvas id="myCanvas" width=800 height=500></canvas>
-<input type="button" id="ihn-demo-park" value="停车"><input type="text" id="ihn-demo-park-id" value="车位号" style="margin-right: 20px;" />
-<input type="button" id="ihn-demo-unpark" value="取车"><input type="text" id="ihn-demo-unpark-id" value="车位号" />
