+++ IHNServer/WebRoot/scripts/app/property/templates/ihn/footer-template.html
@@ -1,1 +1,1 @@
-<span>This is footer</span>
+<span style="display:block;text-align:center;">IHN is pretty sweet, it is produced by IHN</span>
