--- IHNServer/WebRoot/scripts/app/property/templates/ihn/header-template.html
@@ -1,2 +1,0 @@
-<span>This is header</span>
-<span class="logout" style="float: right; margin-right: 10px;"><a>Logout</a></span>
