+++ IHNServer/WebRoot/scripts/app/property/templates/ihn/maincontent-template.html
@@ -1,1 +1,9 @@
-<span>This is main content</span>
+<div style="margin:5px 5px 5px 5px;">
+<input type="button" id="ihn-demo-park" value="停车"><input type="text" id="ihn-demo-park-id" value="车位号" style="margin-right: 20px;" />
+<input type="button" id="ihn-demo-unpark" value="取车"><input type="text" id="ihn-demo-unpark-id" value="车位号" />
+</div>
+
+<div style="margin:5px 5px 5px 5px;">
+<canvas id="myCanvas" width=800 height=500></canvas>
+</div>
+
