--- IHNServer/WebRoot/scripts/app/property/templates/ihn/sidemenu-template.html
@@ -1,6 +1,0 @@
-<span>This is side menu</span>
-<ul>
-  <% _.each(menus, function(menu) { %>
-      <li><a href="<%= menu.link %>"><%= menu.name %></a></li>
-  <% }); %>
-</ul>
