+++ IHNServer/pom.xml
@@ -110,7 +110,12 @@
     <artifactId>jersey-multipart</artifactId>
     <version>1.18</version>
   </dependency>
- 
+   <dependency>
+            <groupId>com.sun.jersey</groupId>
+            <artifactId>jersey-bundle</artifactId>
+            <version>1.18</version>
+        </dependency>
+        
 	<dependency>
 		<groupId>org.eclipse.jetty</groupId>
 		<artifactId>jetty-server</artifactId>
